backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.0
openpyxl==3.1.5
Pillow==10.4.0
python-multipart==0.0.9
python-dotenv==1.0.1
